pyproject.toml
[project]
name = "agentguard"
version = "0.3.0"
description = "Multi-Agent Security & Governance Platform for Enterprise AI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["ai-safety", "agent-security", "guardrails", "llm", "prompt-injection"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-exporter-otlp>=1.24.0",
    "azure-ai-contentsafety>=1.0.0",
    "azure-ai-textanalytics>=5.3.0",
    "azure-identity>=1.16.0",
    "python-dotenv>=1.0.0",
    "sqlparse>=0.5.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "networkx>=3.6.1",
    "river>=0.23.0",
]

[project.urls]
Homepage = "https://github.com/wildcraft958/AgentGuard_NamoFans"
Repository = "https://github.com/wildcraft958/AgentGuard_NamoFans"
Issues = "https://github.com/wildcraft958/AgentGuard_NamoFans/issues"

[project.optional-dependencies]
dashboard = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sse-starlette>=2.0.0",
]
testing = [
    "deepteam>=1.0.6",
    "litellm>=1.40.0",
]
autogen = [
    "autogen-agentchat>=0.4.0",
    "autogen-ext[azure]>=0.4.0",
]
reports = [
    "reportlab>=4.4.10",
    "azure-ai-evaluation>=1.15.3",
]
all = [
    "agentguard[dashboard,testing,autogen,reports]",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
agentguard = "agentguard.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]
